% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_address}
\alias{add_address}
\title{Add a address to the addresses table.}
\usage{
add_address(contact_id, position_id, type, value, db = NULL)
}
\arguments{
\item{contact_id}{Id of contact}

\item{position_id}{id of client}

\item{type}{name of the address type.}

\item{value}{value of address.}

\item{db}{Database file.}
}
\description{
Add a address to the addresses table.
}
