% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_billing_scheme}
\alias{add_billing_scheme}
\title{Add a address to the addresses table.}
\usage{
add_billing_scheme(
  name,
  description,
  rate_period,
  rate_value,
  currency,
  db = NULL
)
}
\arguments{
\item{name}{Name of the scheme}

\item{description}{Description}

\item{rate_period}{unit where rate is applied}

\item{rate_value}{bill per unit}

\item{currency}{currency billing is in.}

\item{db}{Database file.}
}
\description{
Add a address to the addresses table.
}
