% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_client}
\alias{add_client}
\title{Add a client to the clients table.}
\usage{
add_client(name, short_name, long_name, other_info = NULL, db = NULL)
}
\arguments{
\item{name}{Name of the client}

\item{short_name}{Short name of client}

\item{long_name}{Long name of client}

\item{other_info}{Other information.}

\item{db}{Database file.}
}
\description{
Add a client to the clients table.
}
