% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_contact}
\alias{add_contact}
\title{Add a contact to the contacts table.}
\usage{
add_contact(title, firstname, surname, fullname, other_info = NULL, db = NULL)
}
\arguments{
\item{title}{Title (Mr/Ms) of contact}

\item{firstname}{First name}

\item{surname}{Last name/surname}

\item{fullname}{Full name including title, firstname, surname}

\item{other_info}{Other information.}

\item{db}{Database file.}
}
\description{
Add a contact to the contacts table.
}
