% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_invoice}
\alias{add_invoice}
\title{Add a task to the addresses table.}
\usage{
add_invoice(
  issue_date,
  due_period_unit,
  due_period_value,
  sent,
  paid,
  other_info,
  db = NULL
)
}
\arguments{
\item{issue_date}{Date to issue.}

\item{due_period_unit}{Period unit to use.}

\item{due_period_value}{Period length}

\item{sent}{Sent?}

\item{paid}{Paid?}

\item{other_info}{Other info to add}

\item{db}{Database file.}
}
\description{
Add a task to the addresses table.
}
