% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_invoice_task_group_allocation}
\alias{add_invoice_task_group_allocation}
\title{Add a task group to the invoice allocation table.}
\usage{
add_invoice_task_group_allocation(invoice_id, task_group_id, db = NULL)
}
\arguments{
\item{invoice_id}{Invoice ID.}

\item{task_group_id}{Task Group ID.}

\item{db}{Database file.}
}
\description{
Add a task group to the invoice allocation table.
}
