% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_position}
\alias{add_position}
\title{Add a position to the positions table.}
\usage{
add_position(
  contact_id,
  client_id,
  name,
  start_date = NA_character_,
  end_date = NA_character_,
  other_info = NULL,
  db = NULL
)
}
\arguments{
\item{contact_id}{Id of contact}

\item{client_id}{id of client}

\item{name}{name of the position.}

\item{other_info}{Other information.}

\item{db}{Database file.}
}
\description{
Add a position to the positions table.
}
