% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_project}
\alias{add_project}
\title{Add a project to the projects table.}
\usage{
add_project(
  client_id,
  name,
  description,
  billing_scheme_id,
  other_info,
  db = NULL
)
}
\arguments{
\item{client_id}{Client ID}

\item{name}{Name of project}

\item{description}{Description of project}

\item{billing_scheme_id}{ID for billing scheme.}

\item{other_info}{Other info about projects.}

\item{db}{Database file.}
}
\description{
Add a project to the projects table.
}
