% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_task_group}
\alias{add_task_group}
\title{Add a task to the task_groups table.}
\usage{
add_task_group(project_id, name, description, db = NULL)
}
\arguments{
\item{project_id}{Project ID}

\item{name}{Name of task group}

\item{description}{Description of task group}

\item{db}{Database file.}
}
\description{
Add a task to the task_groups table.
}
