% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_task_group_allocation}
\alias{add_task_group_allocation}
\title{Add a task to the addresses table.}
\usage{
add_task_group_allocation(task_id, task_group_id, db = NULL)
}
\arguments{
\item{task_id}{Task ID}

\item{task_group_id}{Group ID.}

\item{db}{Database file.}
}
\description{
Add a task to the addresses table.
}
