% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{convert_other_info}
\alias{convert_other_info}
\title{Convert other info into character form}
\usage{
convert_other_info(other_info)
}
\arguments{
\item{other_info}{List to convert to JSON}
}
\description{
Convert other info into character form
}
