% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_worklog.R
\name{create_worklog}
\alias{create_worklog}
\title{Create a worklog database for storing work information.}
\usage{
create_worklog(path)
}
\description{
Create a worklog database for storing work information.
}
