% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_xlsx_from_db.R
\name{create_xlsx_from_db}
\alias{create_xlsx_from_db}
\title{Create an XLSX from database}
\usage{
create_xlsx_from_db(path, db = NULL)
}
\arguments{
\item{path}{Path to save XLSX to.}

\item{db}{data-base}
}
\description{
Create an XLSX from database
}
