% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{get_unique_id}
\alias{get_unique_id}
\title{Internal function for new ID creation}
\usage{
get_unique_id(db, table)
}
\arguments{
\item{db}{data-base}

\item{table}{table to create ID for}
}
\description{
Internal function for new ID creation
}
